+++ src/client/main.jsx
@@ -17,7 +17,6 @@ const router = createBrowserRouter([
     element: <Root />,
     children: [
       { path: "/", element: <Students /> },
-      { path: "/tasks", element: <Students /> },
       { path: "/login", element: <AuthForm /> },
       { path: "/students", element: <Students />}
     ],
