+++ src/server/api/index.js
@@ -27,4 +27,4 @@ router.use(async (req, res, next) => {
 });
 
 router.use("/auth", require("./auth"));
-router.use("/tasks", require("./tasks"));
+router.use("/students", require("./students"));
